requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-Assets==2.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
sentry-sdk==1.5.12
six==1.16.0
SQLAlchemy==1.4.36
urllib3==1.26.9
webassets==2.0
Werkzeug==2.1.2
zipp==3.8.0
